\name{GMDH2-package}
\alias{GMDH2-package}
\alias{GMDH2-package}
\docType{package}
\title{
Binary Classification via GMDH-Type Neural Network Algorithms
}
\description{
Performs binary classification via Group Method of Data Handling (GMDH) - type neural network algorithms. There exist two main algorithms available in GMDH() and dceGMDH() functions. GMDH() performs classification via GMDH algorithm for a binary response and returns important variables. dceGMDH() performs classification via diverse classifiers ensemble based on GMDH (dce-GMDH) algorithm. Also, the package produces a well-formatted table of descriptives for a binary response. Moreover, it produces confusion matrix, its related statistics and scatter plot (2D and 3D) with classification labels of binary classes to assess the prediction performance. All 'GMDH2' functions are designed for a binary response. See Dag O. and Yozgatligil C. (2016, ISSN:2073-4859) and Kondo T. and Ueno J. (2016, ISSN:1349-4198) for the details of GMDH algorithms.
}
\details{
\tabular{ll}{
Package: \tab GMDH\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}